ansible/roles/webserver/templates/pip_requirements.txt
amqp==1.4.9
amqplib==1.0.2
anybase32==1.1.0
anyjson==0.3.3
apns==2.0.1
APNSWrapper==0.6.1
billiard==3.3.0.23
boto==2.5.2
celery==3.1.25
celery-with-redis==3.0
contextlib2==0.5.5
decorator==3.4.0
Django==1.11.17
django-celery==3.2.2
django-dirtyfields==1.3.1
django-extensions==1.8.1
django-mathfilters==0.4.0
docutils==0.9.1
funcsigs==1.0.2
gnupg==2.0.2
ipython==0.13
kombu==3.0.37
lxml==2.3.2
mock==2.0.0
mongoengine==0.8.7
oauthlib==0.5.1
olefile==0.44
Paste==1.7.5.1
pbr==3.1.1
Pillow==4.2.1
pkg-resources==0.0.0
prettytable==0.7.2
psutil==5.2.2
psycopg2==2.7.3.2
pyasn1==0.1.6
pycrypto==2.6
pymongo==2.8.1
PyPDF2==1.26.0
pysaml2==1.0.1
PySimpleSOAP==1.16
pystache==0.5.3
python-dateutil==1.5
python-keyczar==0.71rc0
python-memcached==1.48
pytz==2017.2
raven==5.24.2
readline==6.2.4.1
redis==2.10.5
reportlab==3.2.0
repoze.who==1.0.18
requests==1.2.0
requests-oauthlib==0.3.3
simplejson==2.4.0
six==1.11.0
soap2py==1.16
South==0.7.5
StringGenerator==0.3.0
suds==0.4
TatSu==4.3.0
typing==3.6.6
ujson==1.33
uWSGI==2.0.18
zope.interface==4.0.5
